requirements.txt
python-whois
typer